Crate zombienet_backchannel

Provides the possibility to coordination between malicious actors and the zombienet test-runner, allowing to reference runtime’s generated values in the test specifications, through a bidirectional message passing implemented as a backchannel.

Structs

BackchannelItem

Broadcaster

ZOMBIENET_BACKCHANNEL

ZombienetBackchannel

Constants

ZOMBIENET